Place filter bar in separate component
